+++ ark/docs/app/playground/page.tsx
@@ -3,7 +3,7 @@
 import { HomeLayout } from "fumadocs-ui/layouts/home"
 import type { ReactNode } from "react"
 import { FloatYourBoat } from "../../components/FloatYourBoat.tsx"
-import { Playground } from "../../components/Playground.tsx"
+import { Playground } from "../../components/playground/Playground.tsx"
 import { baseOptions } from "../layout.config.tsx"
 
 export type LayoutProps = {
