+++ ark/docs/components/playground/tsserver.ts
@@ -1,0 +1,38 @@
+import type * as Monaco from "monaco-editor"
+import { schemaDts } from "../bundles/schema.ts"
+import { typeDts } from "../bundles/type.ts"
+import { utilDts } from "../bundles/util.ts"
+import { defaultPlaygroundCode, editorFileUri } from "./utils.ts"
+
+const configureTypeScript = (monaco: typeof Monaco): void => {
+	const tsDefaultModeConfig = (
+		monaco.languages.typescript.typescriptDefaults as any
+	)._modeConfiguration
+	tsDefaultModeConfig.hovers = false
+	tsDefaultModeConfig.completionItems = false
+
+	monaco.languages.typescript.typescriptDefaults.setCompilerOptions({
+		strict: true,
+		exactOptionalPropertyTypes: true,
+		target: monaco.languages.typescript.ScriptTarget.ESNext,
+		moduleResolution: monaco.languages.typescript.ModuleResolutionKind.NodeJs,
+		allowNonTsExtensions: true
+	})
+
+	monaco.languages.typescript.typescriptDefaults.addExtraLib(utilDts)
+	monaco.languages.typescript.typescriptDefaults.addExtraLib(schemaDts)
+	monaco.languages.typescript.typescriptDefaults.addExtraLib(typeDts)
+}
+
+export const getInitializedTypeScriptService = async (
+	monaco: typeof Monaco,
+	targetUri = monaco.Uri.parse(editorFileUri)
+): Promise<Monaco.languages.typescript.TypeScriptWorker> => {
+	configureTypeScript(monaco)
+
+	if (!monaco.editor.getModel(targetUri))
+		monaco.editor.createModel(defaultPlaygroundCode, "typescript", targetUri)
+
+	const worker = await monaco.languages.typescript.getTypeScriptWorker()
+	return await worker(targetUri)
+}
