+++ ark/docs/components/playground/utils.ts
@@ -1,16 +1,32 @@
-import * as Monaco from "monaco-editor"
+import type * as Monaco from "monaco-editor"
 
-export const recentRequests: Map<string, number> = new Map()
+export const defaultPlaygroundCode = `import { type } from "arktype"
+
+export const MyType = type({
+    name: "string",
+    age: "number"
+})
+
+export const out = MyType({
+    name: "Anders Hejlsberg",
+    age: null
+})
+`
+
+export const editorFileUri = "file:///main.ts"
+
+type RequestMap = Map<string, number>
+
+const duplicateThresholdMs = 50
+
+const recentRequests: RequestMap = new Map()
 
 export const createPositionHash = (
 	model: Monaco.editor.ITextModel,
 	position: Monaco.Position
 ): string => `${model.uri}:${position.lineNumber}:${position.column}`
 
-export const isDuplicateRequest = (
-	positionHash: string,
-	duplicateThresholdMs = 50
-): boolean => {
+export const isDuplicateRequest = (positionHash: string): boolean => {
 	const now = Date.now()
 	const lastRequest = recentRequests.get(positionHash)
 
@@ -18,31 +34,4 @@ export const isDuplicateRequest = (
 
 	recentRequests.set(positionHash, now)
 	return false
-}
-
-export const findNearestTokenBoundary = (
-	model: Monaco.editor.ITextModel,
-	position: Monaco.Position
-): Monaco.Position => {
-	const lineContent = model.getLineContent(position.lineNumber)
-
-	if (position.column > lineContent.length) return position
-
-	let column = position.column
-	const char = lineContent[column - 1]
-
-	if (/\s/.test(char)) {
-		while (column <= lineContent.length && /\s/.test(lineContent[column - 1]))
-			column++
-
-		if (column > lineContent.length)
-			return new Monaco.Position(position.lineNumber, lineContent.length + 1)
-
-		return new Monaco.Position(position.lineNumber, column)
-	}
-
-	const token = model.getWordAtPosition(position)
-	if (token) return new Monaco.Position(position.lineNumber, token.endColumn)
-
-	return position
 }
